services/data-ingestion/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic>=2.6.0,<2.10.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
redis==5.0.1
influxdb-client==1.38.0
paho-mqtt==1.6.1
python-multipart==0.0.6
python-dateutil==2.8.2
python-dotenv==1.0.0
httpx==0.25.2
python-jose[cryptography]==3.3.0
requests==2.31.0
